fix: stripe get fiat amount

+++ src/main/java/com/ndb/auction/dao/oracle/transactions/stripe/StripePresaleDao.java
@@ -35,7 +35,7 @@ public class StripePresaleDao extends BaseOracleDao implements ITransactionDao {
         m.setConfirmedAt(rs.getTimestamp("UPDATED_AT").getTime());
 		m.setStatus(rs.getBoolean("STATUS"));
 		m.setFiatType(rs.getString("FIAT_TYPE"));
-        m.setFiatAmount(rs.getLong("FIAT_AMOUNT"));
+        m.setFiatAmount(rs.getDouble("FIAT_AMOUNT"));
         m.setPaymentMethodId(rs.getString("METHOD_ID"));
         m.setPaymentIntentId(rs.getString("INTENT_ID"));
 		return m;
